// frontend/rl-app/src/app/features/admin/matrices-riesgos/utils/matrices-riesgos-catalogos-data.generated.ts
export const MATRICES_RIESGOS_CATALOGS = {
// ...
  "CONTROL_EFFECTIVENESS": [
    {
      "key": "INEXISTENTE",
      "label": "Inexistente",
      "dbKey": "0",
      "sortOrder": 1,
      "aliases": [],
      "level": 0,
      "percentage": 0
    },
    {
      "key": "INEFECTIVO",
      "label": "Inefectivo",
      "dbKey": "1",
      "sortOrder": 2,
      "aliases": [
        "Es inefectivo"
      ],
      "level": 1,
      "percentage": 0
    },
    {
      "key": "RAZONABLE",
      "label": "Razonable",
      "dbKey": "2",
      "sortOrder": 3,
      "aliases": [],
      "level": 2,
      "percentage": 0.3
    },
    {
      "key": "PARCIALMENTE_EFECTIVO",
      "label": "Parcialmente Efectivo",
      "dbKey": "3",
      "sortOrder": 4,
      "aliases": [],
      "level": 3,
      "percentage": 0.5
    },
    {
      "key": "MODERADO",
      "label": "Moderado",
      "dbKey": "4",
      "sortOrder": 5,
      "aliases": [],
      "level": 4,
      "percentage": 0.85
    },
    {
      "key": "ALTA_EFECTIVIDAD",
      "label": "Alta Efectividad",
      "dbKey": "5",
      "sortOrder": 6,
      "aliases": [],
      "level": 5,
      "percentage": 0.9
    }
  ],
  "CONTROL_AUTOMATION": [
    {
      "key": "AUTOMATIZADO",
      "label": "Automatizado",
      "dbKey": "AUTOMATICO",
      "sortOrder": 1,
      "aliases": [
        "AUTOMATICO"
      ]
    },
    {
      "key": "SEMIAUTOMATIZADO",
      "label": "Semiautomatizado",
      "dbKey": "SEMIAUTOMATICO",
      "sortOrder": 2,
      "aliases": [
        "Semi-Automatizado",
        "SEMIAUTOMATICO"
      ]
    },
    {
      "key": "MANUAL",
      "label": "Manual",
      "dbKey": "MANUAL",
      "sortOrder": 3,
      "aliases": []
    }
  ],
  "MONITORING_RISK_STATUS": [
    {
      "key": "VIGENTE",
      "label": "Vigente",
      "dbKey": "VIGENTE",
      "sortOrder": 1,
      "aliases": []
    },
    {
      "key": "MITIGADO",
      "label": "Mitigado",
      "dbKey": "MITIGADO",
      "sortOrder": 2,
      "aliases": []
    },
    {
      "key": "NUEVO",
      "label": "Nuevo",
      "dbKey": "NUEVO",
      "sortOrder": 3,
      "aliases": []
    }
  ],
  "MONITORING_CONTROL_STATUS": [
    {
      "key": "SE_MANTIENE",
      "label": "Se mantiene",
      "dbKey": "SE_MANTIENE",
      "sortOrder": 1,
      "aliases": []
    },
    {
      "key": "NO_SE_MANTIENE",
      "label": "No se mantiene",
      "dbKey": "NO_SE_MANTIENE",
      "sortOrder": 2,
      "aliases": []
    },
    {
      "key": "REQUIERE_ACTUALIZACIÓN",
      "label": "Requiere actualización",
      "dbKey": "REQUIERE_ACTUALIZACIÓN",
      "sortOrder": 3,
      "aliases": []
    }
  ],
  "MONITORING_EFFECTIVENESS": [
    {
      "key": "INEXISTENTE",
      "label": "Inexistente",
      "dbKey": null,
      "sortOrder": 1,
      "aliases": []
    },
    {
      "key": "INEFECTIVO",
      "label": "Inefectivo",
      "dbKey": null,
      "sortOrder": 2,
      "aliases": [
        "Es inefectivo"
      ]
    },
    {
      "key": "RAZONABLE",
      "label": "Razonable",
      "dbKey": null,
      "sortOrder": 3,
      "aliases": []
    },
    {
      "key": "PARCIALMENTE_EFECTIVO",
      "label": "Parcialmente Efectivo",
      "dbKey": null,
      "sortOrder": 4,
      "aliases": []
    },
    {
      "key": "MODERADO",
      "label": "Moderado",
      "dbKey": null,
      "sortOrder": 5,
      "aliases": []
    },
    {
      "key": "ALTA_EFECTIVIDAD",
      "label": "Alta Efectividad",
      "dbKey": null,
      "sortOrder": 6,
      "aliases": []
    }
  ]
} as const;